refactor: split classes.py into enums/ and features/
