+++ src/xplane_apt_convert/enums/lines.py
@@ -1,0 +1,74 @@
+from ._fallback import FallbackEnum
+
+
+class LineType(FallbackEnum):
+    NONE = 0
+    SOLID_YELLOW = 1
+    BROKEN_YELLOW = 2
+    DOUBLE_SOLID_YELLOW = 3
+    RUNWAY_HOLD = 4  # DOUBLE_BROKEN_YELLOW_WITH_PARALLEL_DOUBLE_SOLID_YELLOW = 4
+    OTHER_HOLD = 5  # BROKEN_YELLOW_WITH_PARALLEL_SOLID_YELLOW = 5
+    ILS_HOLD = 6  # YELLOW_CROSSHATCHED = 6
+    ILS_CRITICAL_CENTERLINE = 7
+    SEPARATED_BROKEN_YELLOW = 8
+    SEPARATED_DOUBLE_BROKEN_YELLOW = 9
+    WIDE_SOLID_YELLOW = 10
+    WIDE_ILS_CRITICAL_CENTERLINE = 11
+    WIDE_RUNWAY_HOLD = 12
+    WIDE_OTHER_HOLD = 13
+    WIDE_ILS_HOLD = 14
+
+    SOLID_YELLOW_WITH_BLACK_BORDER = 51
+    BROKEN_YELLOW_WITH_BLACK_BORDER = 52
+    DOUBLE_SOLID_YELLOW_WITH_BLACK_BORDER = 53
+    RUNWAY_HOLD_WITH_BLACK_BORDER = 54
+    OTHER_HOLD_WITH_BLACK_BORDER = 55
+    ILS_HOLD_WITH_BLACK_BORDER = 56
+    ILS_CRITICAL_CENTERLINE_WITH_BLACK_BORDER = 57
+    SEPARATED_BROKEN_YELLOW_WITH_BLACK_BORDER = 58
+    SEPARATED_DOUBLE_BROKEN_YELLOW_WITH_BLACK_BORDER = 59
+    WIDE_SOLID_YELLOW_WITH_BLACK_BORDER = 60
+    WIDE_ILS_CRITICAL_CENTERLINE_WITH_BLACK_BORDER = 61
+    WIDE_RUNWAY_HOLD_WITH_BLACK_BORDER = 62
+    WIDE_OTHER_HOLD_WITH_BLACK_BORDER = 63
+    WIDE_ILS_HOLD_WITH_BLACK_BORDER = 64
+
+    VERY_WIDE_YELLOW = 19
+
+    SOLID_WHITE = 20
+    CHEQUERED_WHITE = 21
+    BROKEN_WHITE = 22
+    SHORT_BROKEN_WHITE = 23
+    WIDE_SOLID_WHITE = 24
+    WIDE_BROKEN_WHITE = 25
+    SOLID_RED = 30
+    BROKEN_RED = 31
+    WIDE_SOLID_RED = 32
+    SOLID_ORANGE = 40
+    SOLID_BLUE = 41
+    SOLID_GREEN = 42
+
+    SOLID_WHITE_WITH_BLACK_BORDER = 70
+    CHEQUERED_WHITE_WITH_BLACK_BORDER = 71
+    BROKEN_WHITE_WITH_BLACK_BORDER = 72
+    SHORT_BROKEN_WHITE_WITH_BLACK_BORDER = 73
+    WIDE_SOLID_WHITE_WITH_BLACK_BORDER = 74
+    WIDE_BROKEN_WHITE_WITH_BLACK_BORDER = 75
+    SOLID_RED_WITH_BLACK_BORDER = 80
+    BROKEN_RED_WITH_BLACK_BORDER = 81
+    WIDE_SOLID_RED_WITH_BLACK_BORDER = 82
+    SOLID_ORANGE_WITH_BLACK_BORDER = 90
+    SOLID_BLUE_WITH_BLACK_BORDER = 91
+    SOLID_GREEN_WITH_BLACK_BORDER = 92
+
+
+class LineLightingType(FallbackEnum):
+    NONE = 0
+    GREEN_BIDIRECTIONAL_LIGHTS = 101
+    BLUE_OMNIDIRECTIONAL_LIGHTS = 102
+    AMBER_UNIDIRECTIONAL_LIGHTS = 103
+    AMBER_UNIDIRECTIONAL_PULSATING_LIGHTS = 104
+    ALTERNATING_AMBER_GREEN_BIDIRECTIONAL_LIGHTS = 105
+    RED_OMNIDIRECTIONAL_LIGHTS = 106
+    GREEN_UNIDIRECTIONAL_LIGHTS = 107
+    ALTERNATING_AMBER_GREEN_UNIDIRECTIONAL_LIGHTS = 108
