+++ src/xplane_apt_convert/features/_base.py
@@ -1,0 +1,12 @@
+from abc import ABC, abstractmethod
+
+
+class AptFeature(ABC):
+    @staticmethod
+    @abstractmethod
+    def _schema() -> dict:
+        pass
+
+    @abstractmethod
+    def _to_record(self) -> dict:
+        pass
